Added CarInfoScreen UI interface for user cars, and changed Id's type in entities from Int to Long

+++ app/src/main/java/com/example/parkpal/data/entity/ParkingLocationEntity.kt
@@ -24,9 +24,9 @@ import androidx.room.PrimaryKey
 )
 data class ParkingLocationEntity(
     @PrimaryKey(autoGenerate = true)
-    val parkingLocationId: Int = 0,
-    val userId: Int, // Foreign key to User table
-    val carId: Int, // Foreign key to Car table
+    val parkingLocationId: Long = 0,
+    val userId: Long, // Foreign key to User table
+    val carId: Long, // Foreign key to Car table
     val latitude: Double,
     val longitude: Double,
     val timestamp: String
