+++ app/src/main/java/com/example/parkpal/domain/model/Car.kt
@@ -1,8 +1,8 @@
 package com.example.parkpal.domain.model
 
 data class Car(
-    val carId: Int = 0,
-    val userId: Int, // Foreign key to User table
+    val carId: Long = 0,
+    val userId: Long, // Foreign key to User table
     val brand: String,
     val model: String,
     val year: Int,
