+++ app/src/main/java/com/example/parkpal/domain/model/ParkingLocation.kt
@@ -1,9 +1,9 @@
 package com.example.parkpal.domain.model
 
 data class ParkingLocation(
-    val parkingLocationId: Int = 0,
-    val userId: Int, // Foreign key to User table
-    val carId: Int, // Foreign key to Car table
+    val parkingLocationId: Long = 0,
+    val userId: Long, // Foreign key to User table
+    val carId: Long, // Foreign key to Car table
     val latitude: Double,
     val longitude: Double,
     val timestamp: String
